Actually fix rendering when the cell has float size
This may seem a naive approach but it's the one that gives the
best results, and no drawback as far as i can see,

+++ src/screen.h
@@ -82,8 +82,8 @@ private:
 
     QPainter *m_painter;
     struct {
-        qreal cellW;
-        qreal cellH;
+        int cellW;
+        int cellH;
         QFont font;
         tsm_age_t age;
     } m_renderdata;
